refactor: SideBar 정적데이터 분리
