+++ src/data/SideBarMenu.ts
@@ -1,0 +1,30 @@
+const challengeMenuList = [
+  {
+    name: "챌린지 참여 현황",
+    path: "/challenge/dashboard",
+  },
+  {
+    name: "챌린지 정보",
+    path: "/challenge/info",
+  },
+  {
+    name: "챌린지 질문 관리",
+    path: "/challenge/question",
+  },
+  {
+    name: "기능 커스텀",
+    path: "/challenge/custom",
+  },
+];
+const participationMenuList = [
+  {
+    name: "참여자 정보",
+    path: "/participation/info",
+  },
+  {
+    name: "참여자 초대",
+    path: "/participation/participate",
+  },
+];
+
+export { challengeMenuList, participationMenuList };
